+++ src/components/SecretCard.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import { getFileNameFromPath } from "../toolbox";
+import { getFileNameFromPath } from "../util/toolbox";
 
 function SecretCard(props) {
   return (
